# book-network-ui/src/app/services/models/borrowed-books-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BorrowedBooksResponse {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}