# book-network-ui/src/openapi/openapi.json
{
     "openapi": "3.0.1",
     "info": {
          "title": "OpenApi specification - Book networking",
          "description": "OpenApi documentation for Spring Security example",
          "termsOfService": "Term of service",
          "contact": {
               "name": "dicka",
               "url": "sidicka.blogspot.co.id",
               "email": "dickanirwansyah@gmail.com"
          },
          "license": {
               "name": "License name",
               "url": "sidicka.blogspot.co.id"
          },
          "version": "1.0"
     },
     "servers": [
          {
               "url": "http://localhost:8081/api/v1",
               "description": "Local ENV"
          },
          {
               "url": "https://dickanirwansyah.aws.co.id/api/v1",
               "description": "Prod ENV"
          }
     ],
     "security": [
          {
               "bearerAuth": []
          }
     ],
     "paths": {
...
          "/books/save-book": {
               "post": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "saveBook",
                    "requestBody": {
                         "content": {
                              "application/json": {
                                   "schema": {
                                        "$ref": "#/components/schemas/BookRequest"
                                   }
                              }
                         },
                         "required": true
                    },
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/cover/{book-id}": {
               "post": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "uploadBookCoverPicture",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "requestBody": {
                         "content": {
                              "multipart/form-data": {
                                   "schema": {
                                        "required": [
                                             "file"
                                        ],
                                        "type": "object",
                                        "properties": {
                                             "file": {
                                                  "type": "string",
                                                  "format": "binary"
                                             }
                                        }
                                   }
                              }
                         }
                    },
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "object"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/borrowed/{book-id}": {
               "post": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "borrowBook",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
          "/books/shareable/{book-id}": {
               "patch": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "updateShareableStatus",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/borrow/return/{book-id}": {
               "patch": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "returnBook",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/borrow/return/approva/{book-id}": {
               "patch": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "approveReturnBorrowBook",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/archived/{book-id}": {
               "patch": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "updateArchivedStatus",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
          "/books/findbook-byid/{book-id}": {
               "get": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "findBookById",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/BookResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/findall-returned-books": {
               "get": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "findAllReturnedBooks",
                    "parameters": [
                         {
                              "name": "page",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 0
                              }
                         },
                         {
                              "name": "size",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 10
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/PageResponseBorrowedBooksResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/findall-books": {
               "get": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "findAllBooks",
                    "parameters": [
                         {
                              "name": "page",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 0
                              }
                         },
                         {
                              "name": "size",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 10
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/PageResponseBookResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/findall-books-by-owner": {
               "get": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "findAllBooksByOwner",
                    "parameters": [
                         {
                              "name": "page",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 0
                              }
                         },
                         {
                              "name": "size",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 10
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/PageResponseBookResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/books/findall-book-borrowed-books": {
               "get": {
                    "tags": [
                         "Book"
                    ],
                    "operationId": "findAllBorrowedBooks",
                    "parameters": [
                         {
                              "name": "page",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 0
                              }
                         },
                         {
                              "name": "size",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 10
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/PageResponseBorrowedBooksResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
     },
     "components": {
          "schemas": {
...
               "BookRequest": {
                    "required": [
                         "authorName",
                         "isbn",
                         "synopsis",
                         "title"
                    ],
                    "type": "object",
                    "properties": {
                         "id": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "title": {
                              "type": "string"
                         },
                         "authorName": {
                              "type": "string"
                         },
                         "isbn": {
                              "type": "string"
                         },
                         "synopsis": {
                              "type": "string"
                         },
                         "shareable": {
                              "type": "boolean"
                         }
                    }
               },
...
               "BookResponse": {
                    "type": "object",
                    "properties": {
                         "id": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "title": {
                              "type": "string"
                         },
                         "authorName": {
                              "type": "string"
                         },
                         "isbn": {
                              "type": "string"
                         },
                         "synopsis": {
                              "type": "string"
                         },
                         "owner": {
                              "type": "string"
                         },
                         "cover": {
                              "type": "array",
                              "items": {
                                   "type": "string",
                                   "format": "byte"
                              }
                         },
                         "rate": {
                              "type": "number",
                              "format": "double"
                         },
                         "archived": {
                              "type": "boolean"
                         },
                         "shareable": {
                              "type": "boolean"
                         }
                    }
               },
               "BorrowedBooksResponse": {
                    "type": "object",
                    "properties": {
                         "id": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "title": {
                              "type": "string"
                         },
                         "authorName": {
                              "type": "string"
                         },
                         "isbn": {
                              "type": "string"
                         },
                         "synopsis": {
                              "type": "string"
                         },
                         "owner": {
                              "type": "string"
                         },
                         "cover": {
                              "type": "array",
                              "items": {
                                   "type": "string",
                                   "format": "byte"
                              }
                         },
                         "rate": {
                              "type": "number",
                              "format": "double"
                         },
                         "archived": {
                              "type": "boolean"
                         },
                         "shareable": {
                              "type": "boolean"
                         }
                    }
               },
               "PageResponseBorrowedBooksResponse": {
                    "type": "object",
                    "properties": {
                         "content": {
                              "type": "array",
                              "items": {
                                   "$ref": "#/components/schemas/BorrowedBooksResponse"
                              }
                         },
                         "number": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "size": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "totalElements": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "totalPages": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "first": {
                              "type": "boolean"
                         },
                         "last": {
                              "type": "boolean"
                         }
                    }
               },
               "PageResponseBookResponse": {
                    "type": "object",
                    "properties": {
                         "content": {
                              "type": "array",
                              "items": {
                                   "$ref": "#/components/schemas/BookResponse"
                              }
                         },
                         "number": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "size": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "totalElements": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "totalPages": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "first": {
                              "type": "boolean"
                         },
                         "last": {
                              "type": "boolean"
                         }
                    }
               }
          },
          "securitySchemes": {
               "bearerAuth": {
                    "type": "http",
                    "description": "JWT auth description",
                    "in": "header",
                    "scheme": "bearer",
                    "bearerFormat": "JWT"
               }
          }
     }
}